refactor: move all Search screen specific components into same folder
