--- src/components/Search/index.ts
@@ -1,3 +1,0 @@
-export { default as SearchBar } from './Bar'
-export { default as SearchResult } from './Result'
-export { default as SearchResultList } from './Results'
