+++ src/screens/Search/Bar.spec.tsx
